Added a mojo to simply update function code
There are times where the scaffolding of the Lambda function is created elsewhere (e.g. - Terraform), and all one wants to do on the build is update the function code of an existing Lambda function(s). Created an “update-lambda-code” target that allows this. Moved common functions from the DeployLambdaMojo to the AbstractLambdaMojo to support this.