deploy λ function if local code ≠ s3 code
Motivation

Assuming the S3 location is the canonical source for your λ code
it is right to deploy a function only when the code differs.

Previously the decision to update was base on either of three
things:
- function does not exist on AWS Lambda
- function JSON configuration has changed (local ≠ lambda)
- forced update using the <forceUpdate> setting

Changes

λ functions are deployed when the code package's MD5 hash
differs with the S3 object.

For example, say your code.zip file's hash is XXXY and S3's
object s3://yourpackages/code.zip has a hash of XXXZ. Your lambda
functions will be deployed in this case.

If the hashes match then no code changes were made, hence no
deployment is necessary. One could still <forceUpdate> if they
wanted to.

Code Changes

- Introduced outcome enum to determine if local function code
  differs with S3.
- Cleaned up create/update process to not rely on exception
  handing.
- Consistent use of String.format.

Problems

- MD5 hashes have the potential for collisions but it's probably
  unlikely. S3 only computes MD5 checksums for objects anyway. In
  the future we could compute SHA256 checksums and store on the
  objects though.